requirements.txt
alembic==1.4.2
aniso8601==8.0.0
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cognitojwt==1.2.2
ecdsa==0.14.1
Flask==1.1.2
Flask-Cognito==1.15
Flask-Migrate==2.5.3
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4
healthcheck==1.3.3
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
numpy==1.19.1
pandas==1.1.0
pdfminer.six==20200104
pdfplumber==0.5.22
Pillow==7.2.0
psycopg2==2.8.5
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.9.8
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-jose==3.2.0
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
sortedcontainers==2.2.2
SQLAlchemy==1.3.18
unicodecsv==0.14.1
urllib3==1.25.10
Wand==0.6.2
Werkzeug==0.16.1
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
